fix(account): refactor imports for user operation constants in backup confirmation and index files
