+++ packages/app/features/account/backup/index.tsx
@@ -30,7 +30,8 @@ import { COSEECDHAtoXY } from 'app/utils/passkeys'
 import { useSendAccount } from 'app/utils/send-accounts/useSendAccounts'
 import { useSupabase } from 'app/utils/supabase/useSupabase'
 import { throwIf } from 'app/utils/throwIf'
-import { defaultUserOp, useUserOpTransferMutation } from 'app/utils/useUserOpTransferMutation'
+import { useUserOpTransferMutation } from 'app/utils/useUserOpTransferMutation'
+import { defaultUserOp } from 'app/utils/userOpConstants'
 import { useAccountNonce } from 'app/utils/userop'
 import type { UserOperation } from 'permissionless'
 import { useCallback, useState } from 'react'
